Turn off and on lights
Same as curtains, but the lights turn on and off if you click on the lamp and or the light switch! To reposition the light switch, you need to move the box collider position to the area. Global light intensity determines how bright we want the entire room. Using temp assets for now.